Add a few options to the script
Update help section and add a few options to the scripts

+++ fectchxx.pl
@@ -1,7 +1,7 @@
 # Crab imges from site xxgege.net/
 # This is a study project, just for fun.
 # Date: 2018/1/3
-# Last Modified: 2018/1/18
+# Last Modified: 2018/1/19
 # Author: Kyle Li
 
 use strict;
@@ -19,11 +19,12 @@ use File::Spec;
 use Getopt::Long;
 use HTTP::Tiny;
 use threads;
+use threads::shared;
 use utf8;
 
 use Md5Hash;
 
-my %imgdb;
+my %imgdb : shared;
 
 # Defined the clean sub while control + C was called to stop the 
 # script
@@ -45,6 +46,9 @@ $SIG{INT} = sub {
 };
 
 # Remove those incompleted and blank ones
+# Under detached mode the scripts could not clean some unfinished
+# images, as they are die abnormally along with the main thread
+# who was sent INT signal
 sub wanted {
    m/.*\.jpg-.*/i && unlink($_);
    m/.*\.php/i && unlink($_);
@@ -69,12 +73,16 @@ my $help;
 my $rand;
 my $art;
 my $page;
+my $freq;
+my $mode;
 
 my %options = (
    'help'   => \$help,
    'rand'   => \$rand,
    'art=s'  => \$art,
-   'page=i' => \$page
+   'page=i' => \$page,
+   'freq=i' => \$freq,
+   'mode=s' => \$mode
 );
 
 GetOptions(%options);
@@ -93,6 +101,15 @@ $url  = $base . '/' . $art;
 
 if (not defined($page)) {
    $page = 3;
+}
+
+if (not defined($freq)) {
+   $freq = 3;
+}
+
+if (not defined($mode) or
+    $mode !~ /^(detach|join)$/i) {
+   $mode = 'detach'
 }
 
 if (defined($rand)) {
@@ -137,6 +154,9 @@ while (my ($link, $name) = each %info) {
       $ff       = File::Fetch->new(uri => $url);
       $img_file = $to_dir . '/' . $ff->file;
       $thr      = async {
+         # Read the img hash db
+         untie %imgdb;
+         tie %imgdb, 'Md5Hash', 'db/imgdb.yml';
          my $flag = 0;
          RETRY:
             eval { 
@@ -159,11 +179,17 @@ while (my ($link, $name) = each %info) {
                }
             }
        };
-      $thr->detach();
+      
+      if ($mode eq 'detach') {
+         $thr->detach();
+      }
+      else {
+         $thr->join();
+      }
 
       # To avoid aggressive spawning thread which consuming too much resources
       # You could comment this line if you are with a powerful machine
-      sleep(int rand(10));
+      sleep(int rand($freq));
    }
    print "$to_dir\n";
 }
@@ -174,12 +200,19 @@ while (my ($link, $name) = each %info) {
 sub help {
    print encode('gbk',"
    $0 - 抓取脚本, 你懂的 
-   本脚本主要用于抓取图片，网站 http://xxgege.net
+   本脚本主要用于抓取图片，网站 http://xxgege.net, 本程序用多个线程来抓取，本地建立
+   md5样本库，用于去掉重复图片，由于无法在服务器端获取md5值，目前采取下载到本地然后在
+   本地库内遍历，看是否已存在，若是，即删除之。
+
    可以使用 Control + c 来停止脚本
    选项：
       -art [artyz artzp artjq artkt artwm artmt artyd]
       -rand 随机抓取
       -page 抓取页数，默认为3页
+      -freq 抓取频率, 这是一个随机数，默认区间0-3秒, 输入数字
+      -mode 线程模式，[detach|join], 
+            detach : 不关心创建出来的抓取线程的死活, 理论上会快点
+            join   : 会等待并回收创建出来的线程, 理论上慢点
       -help 帮助文档
    
    内容映射：
@@ -192,8 +225,8 @@ sub help {
       artyd - 银荡
    
    例子：
-      perl $0 -art artzp -rand -page 5
-      perl $0 -rand
+      perl $0 -art artzp -rand -page 5 -freq 1
+      perl $0 -rand -mod detach
       perl $0
    ");
    exit;
